+++ cmd/reliably/scan/kubernetes.go
@@ -325,10 +325,11 @@ func scanRun(opts *ScanOptions) (count int, err error) {
 	apiClient := api.NewClientFromHTTP(api.AuthHTTPClient(hostname))
 
 	// Sends the context to Reliably prior executing the command
-	orgID, err := api.CurrentUserOrganizationID(apiClient, hostname)
+	org, err := config.GetCurrentOrgInfo()
 	if err != nil {
 		return
 	}
+	orgID := org.ID
 
 	context = ctx.NewContext()
 	opts.Exec, err = api.SendExecutionContext(apiClient, hostname, orgID, context)